// app/src/generated/models/Csa_labelsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Csa_labelscsa_color = {
  100000000: 'Red',
  100000001: 'Orange',
  100000002: 'Yellow',
  100000003: 'Green',
  100000004: 'Blue',
  100000005: 'Purple',
  100000006: 'Gray'
} as const;
export type Csa_labelscsa_color = keyof typeof Csa_labelscsa_color;
export const Csa_labelsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Csa_labelsstatecode = keyof typeof Csa_labelsstatecode;
export const Csa_labelsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Csa_labelsstatuscode = keyof typeof Csa_labelsstatuscode;

export interface Csa_labelsBase {
  csa_color?: Csa_labelscsa_color;
  csa_labelid: string;
  csa_name?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Csa_labelsstatecode;
  statuscode?: Csa_labelsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Csa_labels extends Csa_labelsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  csa_colorname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}